combine estop modals for desktop and touchscreen
