+++ app/src/organisms/EstopModal/DesktopEstopPressedModal.stories.tsx
@@ -1,18 +1,36 @@
 import * as React from 'react'
-import { DesktopEstopPressedModal } from './DesktopEstopPressedModal'
+import { Provider } from 'react-redux'
+import { createStore } from 'redux'
+
+import { configReducer } from '../../redux/config/reducer'
+import { EstopPressedModal } from '.'
+
+import type { Store } from 'redux'
 import type { Story, Meta } from '@storybook/react'
 
 export default {
-  title: 'App/organisms/DesktopEstopPressedModal',
-  component: DesktopEstopPressedModal,
+  title: 'App/organisms/EstopPressedModal',
+  component: EstopPressedModal,
 } as Meta
 
-const Template: Story<
-  React.ComponentProps<typeof DesktopEstopPressedModal>
-> = args => <DesktopEstopPressedModal {...args} />
+const dummyConfig = {
+  config: {
+    isOnDevice: false,
+  },
+} as any
 
-export const Primary = Template.bind({})
-Primary.args = {
+const store: Store<any> = createStore(configReducer, dummyConfig)
+
+const Template: Story<
+  React.ComponentProps<typeof EstopPressedModal>
+> = args => (
+  <Provider store={store}>
+    <EstopPressedModal {...args} />
+  </Provider>
+)
+
+export const EstopPressed = Template.bind({})
+EstopPressed.args = {
   isActiveRun: true,
   isEngaged: true,
 }
