+++ app/src/organisms/EstopModal/TouchscreenEstopPressedModal.stories.tsx
@@ -3,34 +3,35 @@ import { Provider } from 'react-redux'
 import { createStore } from 'redux'
 
 import { configReducer } from '../../redux/config/reducer'
-import { EstopMissingModal } from '.'
+import { EstopPressedModal } from '.'
 
 import type { Store } from 'redux'
 import type { Story, Meta } from '@storybook/react'
 
 export default {
-  title: 'App/organisms/EstopMissingModal',
-  component: EstopMissingModal,
+  title: 'ODD/organisms/EstopPressedModal',
+  component: EstopPressedModal,
 } as Meta
 
 const dummyConfig = {
   config: {
-    isOnDevice: false,
+    isOnDevice: true,
   },
 } as any
 
 const store: Store<any> = createStore(configReducer, dummyConfig)
 
 const Template: Story<
-  React.ComponentProps<typeof EstopMissingModal>
+  React.ComponentProps<typeof EstopPressedModal>
 > = args => (
   <Provider store={store}>
-    <EstopMissingModal {...args} />
+    <EstopPressedModal {...args} />
   </Provider>
 )
 
-export const EstopMissing = Template.bind({})
-EstopMissing.args = {
+export const EstopPressed = Template.bind({})
+EstopPressed.args = {
   isActiveRun: true,
-  robotName: 'Flexy',
+  isEngaged: true,
+  closeModal: () => {},
 }
